Create cipher (1).sh
A script that runs a shift substitution cipher. The cipher uses a keyword entered by the
user that is used to create a monoalphabetic mapping of plaintext to ciphertext. e script should be able to encrypt a text file passed into it as a parameter/argument and write
the resulting ciphertext to a (new) file.